consistant formatting + kotlin 1.6.0
