+++ api/src/main/kotlin/ca/vikelabs/maps/Application.kt
@@ -6,8 +6,6 @@ import org.http4k.contract.openapi.ApiInfo
 import org.http4k.contract.openapi.v3.OpenApi3
 import org.http4k.core.HttpHandler
 import org.http4k.format.Jackson
-import org.http4k.server.SunHttp
-import org.http4k.server.asServer
 
 val application: HttpHandler = contract {
     renderer = OpenApi3(
@@ -16,11 +14,4 @@ val application: HttpHandler = contract {
     )
     descriptionPath = "/"
     routes += ping()
-}
-
-fun main() {
-    val server = application.asServer(SunHttp())
-    server.start()
-    println("Server started and running on http://localhost:${server.port()}.")
-    server.block()
 }
