--- api/src/main/kotlin/ca/vikelabs/maps/opensteetmaps/MapData.kt
@@ -1,3 +1,0 @@
-package ca.vikelabs.maps.opensteetmaps
-
-interface MapData
