+++ api/src/main/kotlin/ca/vikelabs/maps/opensteetmaps/types/OverpassOsm3s.kt
@@ -1,6 +1,5 @@
 package ca.vikelabs.maps.opensteetmaps.types
 
-
 import com.fasterxml.jackson.annotation.JsonProperty
 
 data class OverpassOsm3s(
